requirement.txt
pandas 
numpy
pymysql
mysql-connector-python
scikit-learn
python-dotenv
sqlalchemy

#-e .